feat: create invoice resource, form, table, pdf

+++ database/migrations/2025_09_19_140718_create_services_table.php
@@ -15,7 +15,7 @@ return new class extends Migration
             $table->id();
             $table->string('service_name');
             $table->string('unit')->nullable();
-            $table->decimal('price', 15, 2)->default(0);
+            $table->integer('price')->default(0);
             $table->timestamps();
         });
     }
